stack/requirements.txt
paho-mqtt
influxdb